# swagger-output.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ethers HD Wallet and Token Transfer Example",
    "version": "1.0.0",
    "description": "This REST API demonstrates how to use the Ethers library to interact with Ethereum using Infura. The API showcases various functionalities, including creating an HD wallet, fetching Ethereum and token balances, sending Ethereum and tokens, estimating gas prices, and more"
  },
  "servers": [
    {
      "url": "http://localhost:5000/",
      "description": "local server"
    },
    {
      "url": "http://localhost:3000/",
      "description": "docker server"
    }
  ],
...
  "paths": {
...
    "/getUser/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "get specific user",
        "description": "",
        "operationId": "getUsers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User id"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/notFound"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/notFound"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/getAllUsers": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "get all users",
        "description": "",
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page number to be enquiry",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "quantity of users object per page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "filter users with there role",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users successfully obtained.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/users"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/users"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/updateUser/{id}": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "update user by id",
        "description": "",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User id"
          }
        ],
        "responses": {
          "200": {
            "description": "Users successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/notFound"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/notFound"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/auth"
              }
            }
          }
        }
      }
    },
    "/deleteUser/{id}": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "delete user by id",
        "description": "",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User id"
          }
        ],
        "responses": {
          "200": {
            "description": "Users successfully deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Document with 64ecbb91f981e92a17e68cb2 has been deleted.."
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Document with 64ecbb91f981e92a17e68cb2 has been deleted.."
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/notFound"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/notFound"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/getTransaction/{txnHash}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "Get specific transactions by transaction hash",
        "description": "",
        "operationId": "getTransactionByTx",
        "parameters": [
          {
            "name": "txnHash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "transaction hash"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction successfully obtained.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/transaction"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/transaction"
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/notFound"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/notFound"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/getTransactions": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "Get all transactions of logged-in user",
        "description": "",
        "operationId": "getTransactions",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page number to be enquiry",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "quantity of transactions object per page",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transactions successfully obtained.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/transaction"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/transaction"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/getAllTransaction": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "Get all transactions",
        "description": "",
        "operationId": "getAllTransaction",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page number to be enquiry",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "quantity of transactions object per page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "filter transactions with user Id",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transactions successfully obtained.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/transaction"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/transaction"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "auth": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "user1"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        },
        "xml": {
          "name": "auth"
        }
      },
...
      "user": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "user1"
          },
          "role": {
            "type": "string",
            "example": "user"
          },
          "address": {
            "type": "string",
            "example": "0x663f53bE3B2aBfD8F5Fdc26D6f86bD429a6df588"
          },
          "id": {
            "type": "string",
            "example": "64ecbb91f981e92a17e68cb2"
          }
        },
        "xml": {
          "name": "user"
        }
      },
      "users": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "xml": {
                "name": "user"
              },
              "$ref": "#/components/schemas/user"
            }
          },
          "totalItems": {
            "type": "number",
            "example": 1
          },
          "offset": {
            "type": "number",
            "example": 0
          },
          "pageSize": {
            "type": "number",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "example": 1
          },
          "currentPage": {
            "type": "number",
            "example": 1
          },
          "pagingCounter": {
            "type": "number",
            "example": 1
          },
          "hasPrevPage": {
            "type": "boolean",
            "example": false
          },
          "hasNextPage": {
            "type": "boolean",
            "example": false
          },
          "prevPage": {},
          "nextPage": {}
        },
        "xml": {
          "name": "users"
        }
      },
...
      "transaction": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "64dff3de85d309f51b53f9f5"
          },
          "isTokenTx": {
            "type": "boolean",
            "example": true
          },
          "value": {
            "type": "number",
            "example": 1
          },
          "toAddress": {
            "type": "string",
            "example": "0xF4623e04B96dFa4fc2267ac1217FC86966f56752"
          },
          "blockNumber": {
            "type": "number",
            "example": 4118919
          },
          "txnHash": {
            "type": "string",
            "example": "0x451aa55fbeffbeacb2c38ea1a0b2524419c0fb6d3e1e86d3ee7b41a533de3b25"
          },
          "txnFee": {
            "type": "number",
            "example": 97294588618500
          },
          "gasPrice": {
            "type": "number",
            "example": 1425561738
          },
          "method": {
            "type": "string",
            "example": "Transfer"
          },
          "id": {
            "type": "string",
            "example": "64e1391d9b18a56af76e0f11"
          }
        },
        "xml": {
          "name": "transaction"
        }
      },
...
      "notFound": {
        "type": "object",
        "properties": {
          "error": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "Not found."
          },
          "stack": {
            "type": "string",
            "example": ""
          }
        },
        "xml": {
          "name": "notFound"
        }
      },
      "error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "number",
            "example": 500
          },
          "message": {
            "type": "string",
            "example": "Something went wrong"
          },
          "stack": {
            "type": "string",
            "example": ""
          }
        },
        "xml": {
          "name": "error"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}